CCDM: move IndexHtmlRequestListener section to application-shell article
